renamed users to menu and few changes to error
